@@ -1,9 +1,15 @@
 ﻿@using NewStore.Application.Services.Products.Queris.GetProduct
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@addTagHelper *, LazZiya.TagHelpers
 @model ResultGetProductDto
 @{
     ViewData["Title"] = "Index";
-    Layout = "~/Views/Shared/_Layout.cshtml";
+     Layout = "~/Views/Shared/_Layout.cshtml";
+    //Layout = null;
+    int correctPage = 1;
+    int pageSize = 20;
+    int.TryParse(Context.Request.Query["page"], out correctPage);
+    int.TryParse(Context.Request.Query["pageSize"], out pageSize);
 }
 <section class="breadcrumb-bg">
     <div class="theme-container container ">
@@ -92,7 +98,7 @@
         <div class="tab-content shop-content">
             <div id="grid-view" class="tab-pane fade" role="tabpanel">
                 <div class="row">
-                    @foreach(var item in Model.Products)
+                    @foreach (var item in Model.Products)
                     {
                         <div class="col-lg-3 col-md-4 col-sm-6">
                             <div class="product-box">
@@ -102,7 +108,7 @@
                                     <div class="prod-icons">
                                         <a href="#" class="fa fa-heart"></a>
                                         <a href="#" class="fa fa-shopping-basket"></a>
-                                        <a asp-action="Details" asp-route-productId="@item.Id"  data-toggle="modal" class="fa fa-expand"></a>
+                                        <a asp-action="Details" asp-route-productId="@item.Id" data-toggle="modal" class="fa fa-expand"></a>
                                     </div>
                                 </div>
                                 <div class="product-caption">
@@ -142,7 +148,7 @@
 
             <div id="list-view" class="tab-pane fade active in" role="tabpanel">
                 <div class="row">
-                    @foreach(var item in Model.Products)
+                    @foreach (var item in Model.Products)
                     {
                         <div class="col-md-6 col-sm-12">
                             <div class="deal-item list-view">
@@ -180,21 +186,33 @@
                 </div>
 
                 <div class="pagination-wrap">
-                    <ul class="pagintn">
-                        <li><a href="#">01</a></li>
-                        <li><a href="#">02</a></li>
-                        <li><a href="#">03</a></li>
-                        <li><a href="#">04</a></li>
-                        <li><a href="#">05</a></li>
-                        <li><a href="#">...</a></li>
-                        <li><a href="#">15</a></li>
-                    </ul>
+                    @* <ul class="pagintn">
+
+                        @for (UInt16 i = 1; i <= Model.RowsCount % 5; i++)
+                        {
+                            <li><a asp-route-page="@i">@i</a></li>
+                        }
+                        @if (Model.RowsCount > 5)
+                        {
+                            <li>...</li>
+                            @for (UInt32 i = Model.RowsCount; i >= Model.RowsCount - 5 && i > Model.RowsCount % 5; i--)
+                            {
+                                <li><a asp-route-page="@i">@i</a></li>
+                            }
+                        }
+                    </ul> *@
+                    <paging total-records="@(int)Model.RowsCount"
+                            page-no="@correctPage == 0 ? 1 : correctPage"
+                            page-size="@pageSize == 0 ? 1 : pageSize"
+                            query-string-key-page-no="page"
+                            query-string-key-page-size="pageSize">
+                    </paging>
                 </div>
             </div>
         </div>
     </div>
 </section>
 
-@section Scripts{
+@section Scripts {
     <script src="~/assets/js/theme.js"></script>
 }
